requirements.txt
pyyaml>=6.0
numpy>=2.0
pandas>=2.2
tqdm>=4.67
python-dotenv>=1.0

torch>=2.11.0
transformers>=5.5.4
accelerate>=1.10.0
safetensors>=0.6.0
sentencepiece>=0.2.0
einops>=0.8.0

qdrant-client>=1.15.0

Pillow>=11.0.0
opencv-python>=4.11.0
torchvision>=0.24.0

easyocr>=1.7.2
openai-whisper>=20240930
spacy>=3.8.0

